requirements.txt
asgiref==3.8.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-cors-headers==4.3.1
django-environ==0.11.2
django-js-asset==2.2.0
django-tinymce==4.0.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
google-auth==2.29.0
idna==3.7
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyJWT==2.8.0
requests==2.31.0
rsa==4.9
sqlparse==0.5.0
urllib3==2.2.1
